requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
sentence-transformers
torch
numpy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
